[SCRUM-294] Rename: NoMoneyChallenge -> PeerChallenge로 이름 변경
